requirements.txt
Flask==3.1.0
langchain==0.3.14
langchain_community==0.3.14
langchain_openai==0.3.1
python-dotenv==1.0.1
Werkzeug==3.1.3
#gunicorn==20.1.0